포스팅 수 1이상이면 삭제 불가능, 블로그 간 이동시 authUser 대신 blog.userID로 변경, authUser와 blog.userID가 다르면 블로그 관리 권한 없음

+++ src/main/java/jblog/controller/CategoryController.java
@@ -24,8 +24,13 @@ public class CategoryController {
     public Map<String, Object> categoryDelete(
         @PathVariable("no") Long no) {
         Map<String, Object> map = new HashMap<String, Object>();
-        map.put("result", "success");
-        map.put("data", categoryService.remove(no));
+        if (categoryService.getByNo(no).getPosting() > 0) {
+            map.put("result", "false");
+        }
+        else {
+            map.put("result", "success");
+            map.put("data", categoryService.remove(no));
+        }
         return map;
     }
 
@@ -35,7 +40,7 @@ public class CategoryController {
         @ModelAttribute CategoryVo categoryVo,
         @PathVariable("id") String id,
         Model model) {
-        BlogVo blogVo = blogService.getByID(id);
+        BlogVo blogVo = blogService.getBlogVoByID(id);
         categoryVo.setBlogNo(blogVo.getNo());
         categoryVo.setPosting(0L);
         model.addAttribute("blogVo", blogVo);
